final-project/contrastive-learning/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "contrastive-learning-imbalanced"
version = "1.0.0"
description = "Contrastive Learning for Handling Imbalanced Data (CIFAR-10-LT and Credit Card Fraud Detection)"
authors = [{name = "ML Researcher", email = "researcher@example.com"}]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "kaggle>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
